Move HipsSurveyProperties and corresponding collection class to the same file
