+++ hips/tiles/__init__.py
@@ -1,5 +1,4 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Classes and functions to manage HiPS tiles."""
 from .tile import *
-from .description import *
 from .surveys import *
